Mostly Complete Cake Recipe Page
Made most of the styling of the Red Velvet Page to match the Homepage.
Colors and images were added to enhance visual appeal as well as the
use of divs to make the content less vertically-organized to add visual
interest. Added many classes for the ease of application for the other
recipe pages that will be made.